graphql_framework_test_app/pyproject.toml
[tool.poetry]
name = "graphql_framework_test_app"
version = "0.1.0"
description = ""
authors = ["Ellis Percival <flyte@failcode.co.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.9"
django-graphql-framework = { path = ".." }
djangorestframework = "^3.11.0"
django-cors-headers = "^3.4.0"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"
pylint-django = "^2.2.0"
mypy = "^0.782"
behave = "^1.2.6"
behave-django = "^1.4.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
